+++ riverpie_flutter/lib/src/scope.dart
@@ -1,32 +1,22 @@
+// ignore_for_file: invalid_use_of_internal_member
+
 import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';
 import 'package:riverpie/riverpie.dart';
+
 // ignore: implementation_imports
 import 'package:riverpie/src/notifier/base_notifier.dart';
+
 // ignore: implementation_imports
 import 'package:riverpie/src/provider/base_provider.dart';
+
 // ignore: implementation_imports
 import 'package:riverpie/src/provider/override.dart';
 
 /// A wrapper widget around [RiverpieContainer].
 class RiverpieScope extends InheritedWidget implements RiverpieContainer {
-  /// If you are unable to access the [ref] for whatever reason,
-  /// there is a pragmatic solution for that.
-  /// This is considered bad practice and should only be used as a last resort.
-  ///
-  /// Usage:
-  /// RiverpieScope.defaultRef.read(myProvider);
-  static late Ref defaultRef;
-
-  /// Holds all provider states
-  final RiverpieContainer _container;
-
-  /// The provided observer (e.g. for logging)
-  @override
-  RiverpieObserver? get observer => _container.observer;
-
   /// Creates a [RiverpieScope].
   /// The [overrides] are used to override providers with a different value.
   /// The [initialProviders] are used to initialize providers right away.
@@ -47,16 +37,29 @@ class RiverpieScope extends InheritedWidget implements RiverpieContainer {
     defaultRef = this;
   }
 
+  /// If you are unable to access the [ref] for whatever reason,
+  /// there is a pragmatic solution for that.
+  /// This is considered bad practice and should only be used as a last resort.
+  ///
+  /// Usage:
+  /// RiverpieScope.defaultRef.read(myProvider);
+  static late Ref defaultRef;
+
+  /// Holds all provider states
+  final RiverpieContainer _container;
+
+  /// The provided observer (e.g. for logging)
+  @override
+  RiverpieObserver? get observer => _container.observer;
+
   /// Returns the actual value of a [Provider].
   @override
-  // ignore: invalid_use_of_internal_member
   T read<N extends BaseNotifier<T>, T>(BaseProvider<N, T> provider) {
     return _container.read(provider);
   }
 
   /// Returns the notifier of a [NotifierProvider].
   @override
-  // ignore: invalid_use_of_internal_member
   N notifier<N extends BaseNotifier<T>, T>(NotifyableProvider<N, T> provider) {
     return _container.notifier(provider);
   }
@@ -65,13 +68,21 @@ class RiverpieScope extends InheritedWidget implements RiverpieContainer {
   /// This method is used internally without
   /// any [NotifyableProvider] constraints.
   @override
-  // ignore: invalid_use_of_internal_member
   N anyNotifier<N extends BaseNotifier<T>, T>(BaseProvider<N, T> provider) {
     return _container.anyNotifier(provider);
   }
 
   @override
-  // ignore: invalid_use_of_internal_member
+  Emittable<N, E> redux<N extends BaseReduxNotifier<T, E>, T, E extends Object>(
+    ReduxProvider<N, T, E> provider,
+  ) {
+    return Emittable(
+      notifier: _container.anyNotifier(provider),
+      debugOwnerLabel: debugOwnerLabel,
+    );
+  }
+
+  @override
   Stream<NotifierEvent<T>> stream<N extends BaseNotifier<T>, T>(
     BaseProvider<N, T> provider,
   ) {
@@ -90,4 +101,7 @@ class RiverpieScope extends InheritedWidget implements RiverpieContainer {
   bool updateShouldNotify(RiverpieScope oldWidget) {
     return false;
   }
+
+  @override
+  String get debugOwnerLabel => 'RiverpieScope';
 }